float fields

+++ Service/models.py
@@ -87,11 +87,11 @@ class Service(models.Model):
     
     #New Fields added 
     controls_time_slot = models.CharField(choices=TIME_SLOT_INTERVAL_CHOICES, blank= True, null=True,  max_length=100, default='30_Mins')
-    initial_deposit =models.PositiveIntegerField(default=0, blank= True, null=True,)
+    initial_deposit =models.FloatField(default=0, blank= True, null=True,)
     client_can_book = models.CharField(choices=CLIENT_CAN_BOOK_CHOICES, blank= True, null=True, max_length=100, default='Anytime')
     slot_availible_for_online = models.CharField(choices=CONTROLS_TIME_SLOT_CHOICES, blank= True, null=True, max_length=100, default='Anytime_In_The_Future')
 
-    price = models.PositiveIntegerField(default=0)
+    price = models.FloatField(default=0)
     
     is_default = models.BooleanField(default=False)
